pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wiwi"
version = "4.0.0"
description = "Modular AI Assistant with pluggable modules"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Wiwi Team"}
]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["ai", "assistant", "llm", "modular", "chatbot"]

dependencies = [
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
llm = [
    "llama-cpp-python>=0.2.0",
]
stt = [
    "faster-whisper>=1.0.0",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "torchaudio>=2.0.0",
]
tts = [
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "TTS>=0.22.0",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
]
tts-chatterbox = [
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "chatterbox-tts",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
]
discord = [
    "py-cord>=2.3.0",
    "PyNaCl>=1.5.0",  # Для голосовых каналов
    "openwakeword>=0.6.0",  # Wake word detection
    "scipy>=1.10.0",  # Для качественного ресэмплинга аудио
]
all = [
    "wiwi[llm,stt,tts,discord]",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
wiwi = "wiwi.__main__:cli_entry"

[project.urls]
Homepage = "https://github.com/wiwi/wiwi4"
Documentation = "https://github.com/wiwi/wiwi4#readme"
Repository = "https://github.com/wiwi/wiwi4"

[tool.setuptools.packages.find]
where = ["."]
include = ["wiwi*"]

[tool.setuptools.package-data]
wiwi = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
